+++ src/index.js
@@ -125,7 +125,8 @@ import {
   useClipboard,
   useDisclosure,
   useTheme,
-  useColorMode
+  useColorMode,
+  CSSReset
 } from '@chakra-ui/core'
 
 export {
@@ -255,7 +256,8 @@ export {
   useClipboard,
   useDisclosure,
   useTheme,
-  useColorMode
+  useColorMode,
+  CSSReset
 }
 
 // export * from '@chakra-ui/core'
@@ -286,6 +288,8 @@ export { default as Footer } from './components/footer'
 // Group
 export { default as Group } from './components/group'
 export { default as Card } from './components/group/card'
+
+// Fieldset
 export { default as Fieldset } from './components/Fieldset'
 export { default as LeftFieldset } from './components/Fieldset/leftFieldset'
 export { default as RightFieldset } from './components/Fieldset/rightFieldset'
@@ -353,3 +357,8 @@ export { default as CrumbLink } from './components/crumb/crumbLink'
 // post
 export { default as Pictures } from './components/postCard/pictures'
 export { default as PictureBox } from './components/postCard/pictureBox'
+
+// phone
+export { default as SingleAnnouncement } from './components/singleAnnouncement'
+export { default as StatusButton } from './components/singleAnnouncement/statusButton'
+export { default as AnnouncementHeader } from './components/singleAnnouncement/announcementHeader'