+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## viviui 1.5.9
+- 增加手机版的通告系统查看通告的 announcementHeader component
+
+## viviui 1.5.2
+- 添加了手机版的通告系统 singleAnnouncement component
+
 ## viviui 1.4.6
 
 - 优化 Table component